+++ R/tar_make_future.R
@@ -30,7 +30,7 @@ tar_make_future <- function(
   reporter = Sys.getenv("TAR_MAKE_REPORTER", unset = "verbose"),
   workers = 1L,
   callr_function = callr::r,
-  callr_arguments = list()
+  callr_arguments = list(spinner = identical(reporter, "silent"))
 ) {
   assert_package("future")
   assert_script()